# tests/GenAIPlatform.IntegrationTests/Baselines/openapi-v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "GenAIPlatform.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost/"
    }
  ],
  "paths": {
...
    "/api/v1/usage": {
      "get": {
        "tags": [
          "v1"
        ],
        "summary": "Get aggregated AI request usage metrics (tokens, cost, latency) filtered by optional dimensions.",
        "operationId": "GetUsage",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "model",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageSummary"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evaluations/runs": {
      "post": {
        "tags": [
          "v1"
        ],
        "summary": "Start a new evaluation run against the configured dataset.",
        "operationId": "StartEvaluationRun",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartEvaluationRunHttpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationRunResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evaluations/runs/{runId}": {
      "get": {
        "tags": [
          "v1"
        ],
        "summary": "Get the full result (cases + checks) of a previously completed evaluation run.",
        "operationId": "GetEvaluationRun",
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationRunResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evaluations/runs/{runId}/summary": {
      "get": {
        "tags": [
          "v1"
        ],
        "summary": "Get a compact summary (counts and headline metrics) of a previously completed evaluation run.",
        "operationId": "GetEvaluationRunSummary",
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationRunSummary"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "EvaluationCaseResult": {
        "required": [
          "caseId",
          "name",
          "status",
          "answer",
          "retrievedCount",
          "retrievalHit",
          "latency",
          "estimatedCost",
          "costCurrency",
          "errorCode",
          "errorMessage",
          "checks"
        ],
        "type": "object",
        "properties": {
          "caseId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "answer": {
            "type": [
              "null",
              "string"
            ]
          },
          "retrievedCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "retrievalHit": {
            "type": "boolean"
          },
          "latency": {
            "pattern": "^-?(\\d+\\.)?\\d{2}:\\d{2}:\\d{2}(\\.\\d{1,7})?$",
            "type": "string"
          },
          "estimatedCost": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "costCurrency": {
            "type": "string"
          },
          "errorCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "errorMessage": {
            "type": [
              "null",
              "string"
            ]
          },
          "checks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EvaluationCheckResult"
            }
          }
        }
      },
      "EvaluationCheckResult": {
        "required": [
          "type",
          "passed",
          "message"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "passed": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "EvaluationFailedCase": {
        "required": [
          "caseId",
          "name",
          "status",
          "errorCode",
          "checks"
        ],
        "type": "object",
        "properties": {
          "caseId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "errorCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "checks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EvaluationCheckResult"
            }
          }
        }
      },
      "EvaluationRunResult": {
        "required": [
          "runId",
          "datasetVersion",
          "runnerVersion",
          "promptVersion",
          "model",
          "modelSettings",
          "retrievalConfiguration",
          "status",
          "startedAtUtc",
          "completedAtUtc",
          "cases"
        ],
        "type": "object",
        "properties": {
          "runId": {
            "type": "string",
            "format": "uuid"
          },
          "datasetVersion": {
            "type": "string"
          },
          "runnerVersion": {
            "type": "string"
          },
          "promptVersion": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "modelSettings": {
            "type": "string"
          },
          "retrievalConfiguration": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "startedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "completedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "cases": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EvaluationCaseResult"
            }
          }
        }
      },
      "EvaluationRunSummary": {
        "required": [
          "runId",
          "datasetVersion",
          "status",
          "totalCases",
          "passedCases",
          "failedCaseCount",
          "retrievalHitRate",
          "averageLatencyMs",
          "averageCost",
          "costCurrency",
          "failedCases"
        ],
        "type": "object",
        "properties": {
          "runId": {
            "type": "string",
            "format": "uuid"
          },
          "datasetVersion": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "totalCases": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "passedCases": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "failedCaseCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "retrievalHitRate": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "averageLatencyMs": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "averageCost": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "costCurrency": {
            "type": "string"
          },
          "failedCases": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EvaluationFailedCase"
            }
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "StartEvaluationRunHttpRequest": {
        "required": [
          "datasetVersion",
          "model",
          "temperature",
          "maxOutputTokens",
          "topK",
          "minSimilarityScore",
          "correlationId"
        ],
        "type": "object",
        "properties": {
          "datasetVersion": {
            "type": [
              "null",
              "string"
            ]
          },
          "model": {
            "type": [
              "null",
              "string"
            ]
          },
          "temperature": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "maxOutputTokens": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "topK": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "minSimilarityScore": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "correlationId": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "UsageSummary": {
        "required": [
          "requests",
          "inputTokens",
          "outputTokens",
          "embeddingTokens",
          "estimatedCost",
          "currency"
        ],
        "type": "object",
        "properties": {
          "requests": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "inputTokens": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "outputTokens": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "embeddingTokens": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "estimatedCost": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "currency": {
            "type": "string"
          }
        }
      }
    }
  },
...
}